backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioimaplib==1.0.1
aioredis==1.3.1
aiosignal==1.3.2
aiosmtplib==3.0.1
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
black==25.1.0
blessed==1.20.0
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==3.0.2
Django==4.2.20
django-cors-headers==4.4.0
django-environ>=0.4.5
django-filter==24.2
django-picklefield==3.3
django-q2==1.7.6
djangorestframework==3.15.2
dnspython==2.7.0
email-validator==2.1.0.post1
filelock==3.18.0
flake8==7.2.0
frozenlist==1.6.0
fsspec==2025.3.2
grpcio==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hiredis==3.1.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
imap-tools==1.10.0
incremental==24.7.2
isort==6.0.1
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pandas==2.0.3
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.7
portalocker==2.10.1
prometheus-client==0.19.0
propcache==0.3.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.3.2
pyOpenSSL==25.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
qdrant-client==1.13.3
redis>=4.6.0
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.13.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
structlog>=21.1.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.19.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.41.0
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.0
zope.interface==7.2
django-prometheus==2.3.0
django-redis>=5.0.0

# API Schema & Docs
drf-spectacular==0.27.2

# Forms
django-crispy-forms==2.1
crispy-bootstrap5==2023.10

django-allauth==0.61.1
django-celery-beat==2.6.0
whitenoise==6.6.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
qrcode[pil]==7.4.2
argon2-cffi==23.1.0
beautifulsoup4==4.12.3

# Testing
pytest>=7.0.0

groq==0.9.0

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
hf_xet==1.0.3

sentry-sdk[django]==1.39.1
django-environ==0.11.2
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1
flower==2.0.1
html2text==2025.4.15
tiktoken==0.5.2
django-guid==3.3.1
markdownify==1.1.0